+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Blocks, missing property
 - Info, missing properties and fix deserialization of a few others
 - ColoredInfo, renaming to ColoredCoinInfo and fix deserialization
+- BlockHeader TransactionCount, TransactionFees refactoring and serialization
 
 
 ## [0.13.0] - 2018-02-18